Add automod cog and create a command to add pattern-based exceptions
