+++ mp2i/static/text/exceptions.json
@@ -1,0 +1,7 @@
+{
+  "bullshit": "nimporte quoi",
+  "immature": "pas mature",
+  "merde": "mince",
+  "retard": "d\u00e9lai",
+  "shit": "mince"
+}
